+++ Makefile
@@ -1,5 +1,2 @@
-html:
-	@go run main.go
-
 development:
 	git rev-parse HEAD > $(CURDIR)/docs/commit.txt
