+++ TODO
@@ -3,6 +3,7 @@
    - Not sure which one currently deploys to heroku, but let's make that deploy to Amazon instead, better understood
    - -web is unnecessary, bring that back home here
    - -api no idea what's up there
+- Move the token generation into 
 - Figure out what's on AWS
 - Get rid of heroku from the process, make that API on AWS
 - More fully separate the API from the UI
