+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/lukemassa/jclubtakeaways-gui
+module github.com/lukemassa/jclubtakeaways
 
 go 1.23
