+++ src/templates/about.html
@@ -27,8 +27,8 @@
         <li>1830s utopian sex cults
         <li>climate refugees
         <li>sock engineering
-        <li>preventing injury in rock climbing
         <li>the history of the Sicilan mafia
+        <li>preventing injury in rock climbing
         <li>constitutional monarchies
         <li>rare earth metals 
         <li>the Toledo war
